task 0 12th commit

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -8,20 +8,19 @@
  *@h: pointer to list
  *Return: number of nodes
  */
-
 size_t print_list(const list_t *h)
 {
 	size_t jm1 = 0;
 
 	while (h)
 	{
-	if (!h->str)
-	printf("[0] (nil)\n");
-	else
-	printf("[%lu] %s\n", h->len, h->str);
+		if (!h->str)
+			printf("[0] (nil)\n");
+		else
+			printf("[%lu] %s\n", (unsigned long)h->len, h->str);
 
-	h = h->next;
-	jm1++;
+		h = h->next;
+		jm1++;
 	}
 	return (jm1);
 }
